Introduced possibility of interactive fit
The main goal of this commit was the implementation of an interactive tool to make the initial guess for the fit paramaters. However, lots of little things appeared along the way. For instance, the parameters that are given in degrees in the configuration file are now converted to radians according to a predefined list also given in the configuration file. The conversion is done only when applying the values to the model, and anywhere else.